pyproject.toml
# pyproject.toml
[project]
name = "energydiff"
version = "2.7.0"
dependencies = [
    "numpy",
    "torch",
    "accelerate",
    "ema-pytorch",
    "einops",
    "matplotlib",
    "pandas",
    "pytorch-lightning",
    "PyYAML",
    "scipy",
    "torchmetrics",
    "tqdm",
    "umap-learn",
    "wandb"
]
[tool.setuptools]
package-dir = {"" = "src"}
